add config field 'replaceVersion'

+++ src/plugin/webpack/VersionPlugin.js
@@ -37,6 +37,7 @@ function VersionPlugin(hashLength, pattern){
             var versions = {};
             var assets = stats.compilation.assets;
             var context = stats.compilation.compiler.context;
+            var config = stats.compilation.compiler.options;
 
             var handerChunk = function(fileName, filePath){
                 if(!pattern || !(pattern && pattern.test(fileName))){
@@ -57,35 +58,37 @@ function VersionPlugin(hashLength, pattern){
                 handerChunk(fileName, filePath);
             }
 
-            var root = path.resolve(__hii__.cwd, 'src/view/');
-            fs.readdir(root, function (err, files) {
-                if(err){
-                    // console.log(err)
-                }else{
-                    files.forEach(function(file){
-                        if(!file.match(/\.(htm|html)/)){
-                            return
-                        }
-                        var filePath = root + '/' + file;
-                        var oldStr = fs.readFileSync(filePath).toString();
-
-                        var newStr = oldStr.replace(/\/([^\/]*?)@(\w+)\.(js|css)/g, function(match, fileName, version, fileExt){
-                            var ver = versions[fileName + '.' + fileExt];
-                            if(ver){
-                                console.log('');
-                                return '/' + fileName + '@' + ver + '.' + fileExt;
-                            }else{
-                                return match;
+            if(config.replaceVersion){
+                var root = path.resolve(__hii__.cwd, config.replaceVersion);
+                fs.readdir(root, function (err, files) {
+                    if(err){
+                        // console.log(err)
+                    }else{
+                        files.forEach(function(file){
+                            if(!file.match(/\.(htm|html)/)){
+                                return
                             }
-                        });
+                            var filePath = root + '/' + file;
+                            var oldStr = fs.readFileSync(filePath).toString();
 
-                        newStr && fs.writeFile(filePath, newStr, function(err){
-                            if(err){ return console.log(err) }
-                            console.log('[debug]', filePath, '版本号替换成功.');
+                            var newStr = oldStr.replace(/\/([^\/]*?)@(\w+)\.(js|css)/g, function(match, fileName, version, fileExt){
+                                var ver = versions[fileName + '.' + fileExt];
+                                if(ver){
+                                    console.log('');
+                                    return '/' + fileName + '@' + ver + '.' + fileExt;
+                                }else{
+                                    return match;
+                                }
+                            });
+
+                            newStr && fs.writeFile(filePath, newStr, function(err){
+                                if(err){ return console.log(err) }
+                                console.log('[debug]', filePath, '版本号替换成功.');
+                            })
                         })
-                    })
-                }
-            });
+                    }
+                });
+            }
 
             // fs.mkdir('ver', function(err){
             mkdirParent(context + '/ver', function(err){
